requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
yfinance>=0.2.36
pandas>=2.0.0
plotly>=5.18.0
langchain>=0.1.0
langchain-openai>=0.0.3
langchain-core>=0.1.10
langchain-community>=0.0.16
openai>=1.4.0
fastapi>=0.111.0
uvicorn>=0.29.0
jinja2>=3.1.3 